Restructured the visualisation module.
