+++ latte/utils/visualisation/images.py
@@ -1,30 +1,19 @@
 """
-Various visualisation helper functions; both for plotting image data, such as homotopies or latent traversals as well as
-for plotting figures of metrics etc.
+Various visualisation helper functions for plotting image data, such as homotopies or latent traversals.
 """
 
-import pathlib
-from typing import Callable, Dict, Optional, List, Any, Union, Tuple
 from itertools import product
+from typing import Callable, Optional, List, Union, Tuple
 
+import matplotlib.pyplot as plt
 import numpy as np
-import pandas as pd
 import torch
+from pythae.models.vae.vae_model import VAE
+from sklearn.preprocessing import StandardScaler
 from torch import nn
 from torch.utils.data import Dataset
-from sklearn.neighbors import KNeighborsRegressor
-from sklearn.preprocessing import StandardScaler
-
-import matplotlib
-import matplotlib.pyplot as plt
-from matplotlib.lines import Line2D
-from mpl_toolkits.mplot3d import art3d
-import seaborn as sns
-
-from pythae.models.vae.vae_model import VAE
 
 from latte.dataset.utils import RandomGenerator
-from latte.models.probabilistic_pca import probabilistic_pca
 from latte.models.vae import utils as vae_utils
 
 left, width = -0.25, 0.5
@@ -592,562 +581,6 @@ def _get_axis_values_2d(
         values = axis_values
 
     return values, qs
-
-
-def _get_lines(
-    X: np.ndarray,
-    A: Optional[np.ndarray] = None,
-    mu: Optional[np.ndarray] = None,
-    A_hat: Optional[np.ndarray] = None,
-    mu_hat: Optional[np.ndarray] = None,
-    dim: int = 3,
-) -> List:
-    """
-    Utility function to get the lines denoting the linear basis of a subspace defined by some projection matrix A and
-    its estimate A_hat.
-    """
-
-    # To plot lines of the appropriate length
-    scale = max([np.max(X[:, j]) - np.min(X[:, j]) for j in range(X.shape[-1])]) / 2
-
-    lines = []
-    params = [(M, v, t) for M, v, t in zip([A, A_hat], [mu, mu_hat], ["truth", "hat"]) if M is not None]
-    for ix, (M, v, t) in enumerate(params):
-        for j in range(M.shape[-1]):
-            v = v if v is not None else np.zeros(dim)
-            if dim == 3:
-                lines.append(
-                    art3d.Line3D(
-                        (-scale * M[0, j] + v[0], scale * M[0, j] + v[0]),
-                        (-scale * M[1, j] + v[1], scale * M[1, j] + v[1]),
-                        (-scale * M[2, j] + v[2], scale * M[2, j] + v[2]),
-                        color="tab:green" if t == "truth" else "tab:orange",
-                        lw=3,
-                    )
-                )
-            elif dim == 2:
-                lines.append(
-                    Line2D(
-                        (-scale * M[0, j] + v[0], scale * M[0, j] + v[0]),
-                        (-scale * M[1, j] + v[1], scale * M[1, j] + v[1]),
-                        lw=2,
-                    )
-                )
-
-    return lines
-
-
-def _plot_points(
-    X: np.ndarray,
-    ax: matplotlib.axes.Axes,
-    dim: int,
-    f: Optional[np.ndarray],
-    color: str,
-    alpha: float,
-    legend_elements: List[Any],
-    legend_names: List[str],
-    legend_colors: List[str],
-    element_name: str,
-) -> None:
-    """
-    Utility function for plotting a set of points onto an axes.
-    It also manages the UI elements which are kept track of.
-    """
-
-    if dim == 2:
-        if f is not None:
-            points = ax.scatter(X[:, 0], X[:, 1], s=10, c=f, alpha=alpha)
-        else:
-            points = ax.scatter(X[:, 0], X[:, 1], s=10, color=color, alpha=alpha)
-    else:
-        if f is not None:
-            points = ax.scatter(X[:, 0], X[:, 1], X[:, 2], s=10, c=f, alpha=alpha)
-        else:
-            points = ax.scatter(X[:, 0], X[:, 1], X[:, 2], s=10, color=color, alpha=alpha)
-    legend_elements.append(points)
-    legend_names.append(element_name)
-    legend_colors.append(color)
-
-
-def synthetic_data(
-    X: np.ndarray,
-    A: Optional[np.ndarray] = None,
-    A_hat: Optional[np.ndarray] = None,
-    f: Optional[np.ndarray] = None,
-    alphas: Dict[str, float] = {"observable": 0.1, "true_projections": 0.2, "estimated_projections": 0.3},
-    file_name: Optional[Union[str, pathlib.Path]] = None,
-) -> None:
-    """
-    Visualises 2D or 3D synthetic data with factors of variation from a nonlinear manifold.
-    Mainly used for data generated with the `nonlinear_manifold` function from the `synthetic` module.
-    Optionally, it also visualises the projections of the points onto 2D or 1D linear subspaces; either the ground-truth
-    projections or the projections on some estimated subspace spanned by the matrices `A` and `A_hat`, respectively.
-    Args:
-        X: A `N x n` array holding the observable data
-        A: A `n x d` matrix projecting X onto the true linear subspace.
-        A_hat: An estimate of the `A` matrix.
-        f: An `N` dimensional vector holding the values of a factor of variation to be color-visualised.
-        alphas: The alpha values for plotting the different points in the figure (the observable ones, the true
-                projections, and the estimated projections).
-                It should contain the keys "observable", "true_projections", "estimated_projections", where the latter
-                two are only needed if also plotting the corresponding points.
-                Useful to configure depending on the number of points plotting.
-        file_name: Optional file name to save the produced figure.
-
-    """
-    dim = X.shape[1]
-    assert dim in [2, 3], "The dimensionality of the data should be 2 or 3 for visualisation."
-
-    fig = plt.figure()
-    if dim == 2:
-        ax = fig.add_subplot(111)
-    else:
-        ax = fig.add_subplot(111, projection="3d")
-    # Construct the lines showing the basis defined by the matrices
-    for line in _get_lines(X=X, A=A, A_hat=A_hat, dim=dim):
-        ax.add_line(line)
-
-    # Useful for the legend on the image
-    legend_elements, legend_names, legend_colors = [], [], []
-
-    _plot_points(
-        X,
-        ax,
-        dim,
-        f,
-        "tab:blue",
-        alphas["observable"],
-        legend_elements,
-        legend_names,
-        legend_colors,
-        "Observable points",
-    )
-
-    if A is not None:
-        true_plane = X @ A @ A.T
-        _plot_points(
-            true_plane,
-            ax,
-            dim,
-            f,
-            "tab:green",
-            alphas["true_projections"],
-            legend_elements,
-            legend_names,
-            legend_colors,
-            "True linear subspace",
-        )
-
-    if A_hat is not None:
-        predicted_plane = X @ A_hat @ A_hat.T
-        _plot_points(
-            predicted_plane,
-            ax,
-            dim,
-            f,
-            "tab:orange",
-            alphas["estimated_projections"],
-            legend_elements,
-            legend_names,
-            legend_colors,
-            "Predicted linear subspace",
-        )
-
-    if dim == 2:
-        ax.set(
-            title=f"{dim}D Gaussian with nonlinear factors of variation",
-            xlabel="x",
-            ylabel="y",
-        )
-    else:
-        ax.set(
-            title=f"{dim}D Gaussian with nonlinear factors of variation",
-            xlabel="x",
-            ylabel="y",
-            zlabel="z" if dim == 3 else None,
-        )
-
-    if f is None:
-        ax.legend(legend_elements, legend_names, labelcolor=legend_colors)
-    else:
-        ax.legend(*legend_elements[0].legend_elements())
-
-    if file_name is not None:
-        fig.savefig(file_name)
-        plt.close()
-    else:
-        plt.show()
-
-
-def probabilistic_pca_data(
-    X: np.ndarray,
-    Z: np.ndarray,
-    A: np.ndarray,
-    mu: np.ndarray,
-    A_hat: Optional[np.ndarray] = None,
-    mu_hat: Optional[np.ndarray] = None,
-    sigma_hat: Optional[float] = None,
-    alphas: Dict[str, float] = {"observable": 0.1, "true_projections": 0.2, "estimated_projections": 0.3},
-    title: Optional[str] = None,
-    file_name: Optional[Union[str, pathlib.Path]] = None,
-) -> None:
-    """
-    Plots 3D observable data coming from the probabilistic PCA generative model, and optionally the subspace spanned by
-    the (estimate of) the loading matrix together with its projections.
-    Mainly used for data generated with the `probabilistic_pca` function from the `synthetic` module.
-    Args:
-        X: A `N x n` matrix of the noisy observations
-        Z: A `N x d` matrix of the generative factors
-        A: The ground-truth `d x d` loading matrix
-        mu: The ground-truth mean of the observable distribution.
-        A_hat: Estimate of `A`
-        mu_hat: Estimate of `mu`
-        sigma_hat: Estimate of `sigma`
-        alphas: The alpha values for plotting the different points in the figure (the observable ones, the true
-                projections, and the estimated projections).
-                It should contain the keys "observable", "true_projections", "estimated_projections", where the latter
-                two are only needed if also plotting the corresponding points.
-                Useful to configure depending on the number of points plotting.
-        title: Optional plot title
-        file_name: Optional file name to save the produced figure.
-
-    """
-    fig = plt.figure()
-    ax = fig.add_subplot(111, projection="3d")
-
-    legend_elements, legend_names, legend_colors = [], [], []
-
-    # Plot the actual noisy observations `X`
-    noise_observations = ax.scatter(X[:, 0], X[:, 1], X[:, 2], s=10, color="tab:blue", alpha=alphas["observable"])
-    legend_elements.append(noise_observations)
-    legend_names.append("Noisy observations")
-    legend_colors.append("tab:blue")
-
-    true_plane = probabilistic_pca.get_observations(Z, A, mu)
-
-    # Construct the lines showing the basis defined by the matrices
-    for line in _get_lines(X=true_plane, A=A, mu=mu, A_hat=A_hat, mu_hat=mu_hat):
-        ax.add_line(line)
-
-    # Plot the projections onto the true subspace
-    true_point_projections = ax.scatter(
-        true_plane[:, 0], true_plane[:, 1], true_plane[:, 2], s=10, color="tab:green", alpha=alphas["true_projections"]
-    )
-    legend_elements.append(true_point_projections)
-    legend_names.append("Projections on the true linear subspace")
-    legend_colors.append("tab:green")
-
-    if A_hat is not None:
-        # Plot the projections onto the estimated subspace
-        Z_hat = probabilistic_pca.get_latent_representations(X, A_hat, mu_hat, sigma_hat)
-        predicted_plane = probabilistic_pca.get_observations(Z_hat, A_hat, mu_hat)
-
-        predicted_point_projections = ax.scatter(
-            predicted_plane[:, 0],
-            predicted_plane[:, 1],
-            predicted_plane[:, 2],
-            s=10,
-            color="tab:orange",
-            alpha=alphas["estimated_projections"],
-        )
-        legend_elements.append(predicted_point_projections)
-        legend_names.append("Projections on the predicted linear subspace")
-        legend_colors.append("tab:orange")
-
-    ax.set(title="Probabilistic PCA" if title is None else title, xlabel="x", ylabel="y", zlabel="z")
-    ax.legend(legend_elements, legend_names, labelcolor=legend_colors)
-
-    if file_name is not None:
-        fig.savefig(file_name)
-        plt.close()
-    else:
-        plt.show()
-
-
-def metrics_boxplot(
-    data: pd.DataFrame,
-    x: str = "Metric",
-    quantity_name: Optional[str] = None,
-    label_rotation: int = 30,
-    file_name: Optional[Union[str, pathlib.Path]] = None,
-) -> None:
-    """
-    Plots the boxplot of the metrics logged in the `data` dataframe.
-    The dataframe should contain a column "Metric" with the name of the metric and a column "Value" which contains the
-    values of the metrics recorded.
-
-    Args:
-        data (pd.DataFrame): The data frame containing the logged metrics.
-        x: The x-axis value of the boxplot.
-        quantity_name (str): The name of the quantity of interest (if one was varied and the values of metrics were
-                             measured at these different points).
-        label_rotation: The rotation of the metric name labels when plotting.
-        file_name (Optional[Union[str, pathlib.Path]], optional): Optional name of the file to save the plot to.
-                                                                  Defaults to None.
-    """
-    n_metrics = len(set(data["Metric"]))
-    fig, ax = plt.subplots(figsize=(max(1.0, 1.0 * n_metrics), 2), dpi=200)
-
-    b = sns.boxplot(data=data, x=x, y="Value", ax=ax)
-    ax.set(
-        title="Metrics" if n_metrics > 1 else list(data["Metric"])[0],
-        xlabel="Metric" if quantity_name is None else quantity_name,
-        ylabel="Value",
-    )
-    ax.set_xticklabels(ax.get_xticklabels(), rotation=label_rotation)
-    b.tick_params(labelsize=6)
-
-    if file_name is not None:
-        fig.savefig(file_name, bbox_inches="tight")
-        plt.close()
-    else:
-        plt.show(bbox_inches="tight")
-
-
-def metric_trend(
-    data: pd.DataFrame,
-    quantity_name: str,
-    title: str = "",
-    file_name: Optional[Union[str, pathlib.Path]] = None,
-) -> None:
-    """
-    Plots the trends in metrics stored in the data frame `data` over different values of a quantity of interest.
-    It assumes that the data frame contains a column names "Metric" with the name of the metrics of interest,
-    a column names "Value" with the measurement of that metric, and a column named "x" which contains the value of
-    the quantity at which the value of the metric was measured.
-
-     Args:
-         data (pd.DataFrame): The data frame containing the data as described above.
-         quantity_name (str): The name of the quantity of interest (which was varied and the values of metrics were
-                              measured at these different points).
-         title (str): The title of the figure
-         file_name (Optional[Union[str, pathlib.Path]], optional): Optional name of the file to save the plot to.
-                                                                   Defaults to None.
-    """
-    fig, ax = plt.subplots(figsize=(4, 2), dpi=200)
-
-    sns.lineplot(data=data, x="x", y="Value", hue="Metric", legend="full", ci="sd", ax=ax)
-    ax.set(xlabel=quantity_name)
-    fig.suptitle(title, fontsize=11)
-    ax.get_legend().remove()
-    # legend = ax.legend(bbox_to_anchor=(1.05, 1), loc=2, borderaxespad=0.0)
-
-    if file_name is not None:
-        fig.savefig(file_name, bbox_inches="tight")
-        # fig.savefig(file_name, bbox_extra_artists=(legend,), bbox_inches="tight")
-        plt.close()
-    else:
-        plt.show()
-
-
-def factor_trend(
-    Z: torch.Tensor,
-    Y: torch.Tensor,
-    target: str,
-    interpolate: bool = True,
-    k: int = 5,
-    N: int = 4000,
-    q: float = 0.05,
-    attribute_names: Optional[List[str]] = None,
-    file_name: Optional[str] = None,
-    rng: RandomGenerator = 42,
-) -> None:
-    """
-    Plots a graph of the values of a set of 1-dimensional variables against a 1-dimensional random variable.
-    It plots the values of the provided set of points or samples `N` points uniformly from the interval containing the
-    1-dimensional variable, and interpolates the values of the set of other variables using `k`-nearest neighbours.
-
-    The function can be used to plot the relationship between a ground-truth factor of variation and a two-dimensional
-    learned representation space or the other way around.
-    Args:
-        Z: The tensor of values of the learned representations of `M` samples.
-        Y: The tensor of values of the true factors of variation of `M` samples.
-        target: The meaning of the target.
-                If "ground-truth", the hue corresponds to the values of the grond-truth factor.
-                In that case, Z.shape[1] must be 1 and Y.shape[1] figures will be produces, one heatmap for each
-                of the columns of `Y`.
-                If "learned", it corresponds to the values of a learned dimension.
-                In that case, Y.shape[1] must be 1 and Z.shape[1] figures will be produces, one heatmap for each
-                of the columns of `Z`.
-        interpolate: Whether to plot the heatmaps as the values of *sampled* points (interpolate = True) based on a
-                     kNN model learned on the true values or just plot the distribution and the heatmap of the original
-                     `M` points provided.
-        k: Number of nearest neighbours to use for interpolation of the values.
-        N: Number of data points to plot.
-        q: The quantile at which to cut off the values of `Y`.
-        attribute_names: A list of length Z.shape[1] (target = "ground-truth") or Y.shape[1] (target = "learned") to
-                         set the titles of all the heatmaps produced.
-        file_name: An optional file name.
-                   If specified, the figure will be saved there, otherwise, it will be displayed.
-        rng: A rando generator for the uniform sampling.
-
-    """
-    assert target in ["ground-truth", "learned"]
-
-    rng = np.random.default_rng(rng)
-
-    if target == "learned":
-        assert Y.shape[1] == 1
-        # Plot Z against Y
-        X = Y
-        T = Z
-
-        xlabel, ylabel = "Factor", "Learned Dimension"
-    else:
-        # Plot Y against Z
-        assert Z.shape[1] == 1
-        X = Z
-        T = Y
-
-        xlabel, ylabel = "Learned Dimension", "Factor"
-
-    fig, ax = plt.subplots(nrows=1, ncols=len(attribute_names), figsize=(2 * len(attribute_names), 2), dpi=120)
-    ax = ax if len(attribute_names) > 1 else [ax]  # in case only a single axes object is returned
-
-    rng = np.random.default_rng(rng)
-
-    for jj in range(len(attribute_names)):
-
-        X_jj, T_jj = (X, T[:, jj]) if target == "ground-truth" else (X[:, [jj]], T)
-
-        if interpolate:
-            regressor = KNeighborsRegressor(n_neighbors=k).fit(X_jj, T_jj)
-
-            # Generate a uniform sample from the range of `X`
-            x = np.sort(
-                rng.uniform(low=-1, high=1, size=(N,))
-                * np.asarray([torch.quantile(X_jj, 1 - q) - torch.quantile(X_jj, q)])
-            )
-            # Compute the values of the variable in `T[:, jj` for the sampled points based on kNN
-            t = regressor.predict(x.reshape(-1, 1))
-
-        else:
-            ixs = rng.choice(len(X), size=min(N, len(X)), replace=False)
-            x = X_jj[ixs]
-            t = T_jj[ixs]
-
-        ax[jj].plot(x, t, label="Latent dimension")
-        ax[jj].set(
-            xlabel=xlabel,
-            ylabel=ylabel,
-            title=attribute_names[jj] if attribute_names is not None else "",
-        )
-        ax[jj].get_xaxis().set_ticks([])
-        ax[jj].get_yaxis().set_ticks([])
-        ax[jj].tick_params(labelsize=6)
-
-    fig.tight_layout()
-
-    if file_name is not None:
-        fig.savefig(file_name, bbox_inches="tight")
-        plt.close()
-    else:
-        plt.show()
-
-
-def factor_heatmap(
-    Z: torch.Tensor,
-    Y: torch.Tensor,
-    target: str,
-    interpolate: bool = True,
-    k: int = 5,
-    N: int = 4000,
-    q: float = 0.05,
-    attribute_names: Optional[List[str]] = None,
-    file_name: Optional[str] = None,
-    rng: RandomGenerator = 42,
-) -> None:
-    """
-    Plots a "jet" heatmap of the values of a set of 1-dimensional variables against a 2-dimensional random variable.
-    It plots the values of the provided set of points or samples `N` points uniformly from the square containing the
-    2-dimensional variable, and interpolates the values of the set of other variables using `k`-nearest neighbours.
-
-    The function can be used to plot the relationship between a ground-truth factor of variation and a two-dimensional
-    learned representation space or the other way around.
-    Args:
-        Z: The tensor of values of the learned representations of `M` samples.
-        Y: The tensor of values of the true factors of variation of `M` samples.
-        target: The meaning of the target.
-                If "ground-truth", the hue corresponds to the values of the grond-truth factor.
-                In that case, Z.shape[1] must be 2 and Y.shape[1] figures will be produces, one heatmap for each
-                of the columns of `Y`.
-                If "learned", it corresponds to the values of a learned dimension.
-                In that case, Y.shape[1] must be 2 and Z.shape[1] figures will be produces, one heatmap for each
-                of the columns of `Z`.
-        interpolate: Whether to plot the heatmaps as the values of *sampled* points (interpolate = True) based on a
-                     kNN model learned on the true values or just plot the distribution and the heatmap of the original
-                     `M` points provided.
-        k: Number of nearest neighbours to use for interpolation of the values.
-        N: Number of data points to plot.
-        q: The quantile at which to cut off the values of `Y`.
-        attribute_names: A list of length Z.shape[1] (target = "ground-truth") or Y.shape[1] (target = "learned") to
-                         set the titles of all the heatmaps produced.
-        file_name: An optional file name.
-                   If specified, the figure will be saved there, otherwise, it will be displayed.
-        rng: A rando generator for the uniform sampling.
-
-    """
-    assert target in ["ground-truth", "learned"]
-
-    if target == "learned":
-        # Plot Z against Y
-        assert Y.shape[1] == 2
-        X = Y
-        T = Z
-
-        axis_name = "Factor"
-    else:
-        # Plot Y against Z
-        assert Z.shape[1] == 2
-        X = Z
-        T = Y
-
-        axis_name = "Learned Dimension"
-
-    fig, ax = plt.subplots(nrows=1, ncols=len(attribute_names), figsize=(2 * len(attribute_names), 2), dpi=120)
-    ax = ax if len(attribute_names) > 1 else [ax]  # in case only a single axes object is returned
-
-    rng = np.random.default_rng(rng)
-
-    for jj in range(len(attribute_names)):
-        X_jj, T_jj = (X, T[:, jj]) if target == "ground-truth" else (X[:, [jj]], T)
-
-        if interpolate:
-            regressor = KNeighborsRegressor(n_neighbors=k).fit(X_jj, T_jj)
-
-            # Generate a uniform sample from the range of `X`
-            x = rng.uniform(low=-1, high=1, size=(N, 2)) * np.asarray(
-                [
-                    torch.quantile(X[:, 0], 1 - q) - torch.quantile(X[:, 0], q),
-                    torch.quantile(X[:, 1], 1 - q) - torch.quantile(X[:, 1], q),
-                ]
-            )
-
-            # Compute the values of the variable in `T[:, jj]` for the sampled points based on kNN
-            t = regressor.predict(x)
-
-        else:
-            ixs = rng.choice(len(X), size=min(N, len(X)), replace=False)
-            x = X_jj[ixs]
-            t = T_jj[ixs]
-
-        ax[jj].scatter(x[:, 0], x[:, 1], c=t, cmap="jet", s=3, label=t)
-        ax[jj].set(
-            xlabel=f"{axis_name} 1",
-            ylabel=f"{axis_name} 2",
-            title=attribute_names[jj] if attribute_names is not None else "",
-        )
-        ax[jj].get_xaxis().set_ticks([])
-        ax[jj].get_yaxis().set_ticks([])
-        ax[jj].tick_params(labelsize=6)
-
-    fig.tight_layout()
-
-    if file_name is not None:
-        fig.savefig(file_name, bbox_inches="tight")
-        plt.close()
-    else:
-        plt.show()
 
 
 def generated_images(
